{{ cookiecutter.repo_name }}/requirements.txt
-e .

## Db
alembic==1.13.1
psycopg2==2.9.9

# Logging
rich==13.3.5
python-json-logger==2.0.7

# Environment
python-dotenv==1.0.0

## Schemas
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
